requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.8.0
requests>=2.28.0
python-dotenv>=1.0.0
gunicorn>=21.0.0 